style: satisfy pre-commit formatting after API move
