+++ src/hibikasu_agent/api/__init__.py
@@ -1,0 +1,1 @@
+"""API package for Hibikasu PRD Reviewer (mock endpoints)."""
